Minor fixes in AWS configuration. Added Db migration
