+++ backend/UserService/src/UserService.API/UserService.API/appsettings.json
@@ -23,8 +23,6 @@
   "AWS": {
     "AvatarBucketName": "${AVATAR_BUCKET_NAME}",
     "AvatarBaseUrl": "${AVATAR_BASE_URL}",
-    "Region": "${AWS_REGION}",
-    "AccessKey": "${AWS_ACCESS_KEY}",
-    "SecretKey": "${AWS_SECRET_KEY}"
+    "Region": "${AWS_REGION}"
   }
 }
